Rename SPUUpdateState back to SPUUserUpdateState
