--- Sparkle/SPUUserUpdateChoice.h
@@ -1,18 +1,0 @@
-//
-//  SPUUserUpdateChoice.h
-//  Sparkle
-//
-//  Created by Mayur Pawashe on 5/31/21.
-//  Copyright © 2021 Sparkle Project. All rights reserved.
-//
-
-#ifndef SPUUserUpdateChoice_h
-#define SPUUserUpdateChoice_h
-
-typedef NS_ENUM(NSInteger, SPUUserUpdateChoice) {
-    SPUUserUpdateChoiceSkip,
-    SPUUserUpdateChoiceInstall,
-    SPUUserUpdateChoiceDismiss,
-};
-
-#endif /* SPUUserUpdateChoice_h */
